Build fix

+++ Telegrom.Database/Model/IdentityState.cs
@@ -6,7 +6,7 @@ namespace Telegrom.Database.Model
     public sealed class IdentityState
     {
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public int IdentityId { get; set; }
+        public long IdentityId { get; set; }
         public string StateName { get; set; }
         public byte[] RowVersion { get; set; }
     }
